tests: fix tests calling themselves
